move method compute_individual_ages_from_biomarker_values from abstract model to abstract multivariate model
