+++ src/leaspy/models/abstract_multivariate_model.py
@@ -1,5 +1,5 @@
 import warnings
-from typing import Optional
+from typing import Optional, Union
 
 import torch
 
@@ -7,7 +7,7 @@ from leaspy.exceptions import LeaspyModelInputError
 from leaspy.io.data.dataset import Dataset
 from leaspy.utils.docs import doc_with_super
 from leaspy.utils.functional import Exp, MatMul
-from leaspy.utils.typing import KwargsType
+from leaspy.utils.typing import DictParams, DictParamsTorch, FeatureType, KwargsType
 from leaspy.utils.weighted_tensor import TensorOrWeightedTensor
 from leaspy.variables.distributions import Normal
 from leaspy.variables.specs import (
@@ -305,3 +305,85 @@ class AbstractMultivariateModel(AbstractModel):
             ].tolist()
 
         return model_settings
+
+    # TODO: unit tests? (functional tests covered by api.estimate)
+    def compute_individual_ages_from_biomarker_values(
+        self,
+        value: Union[float, list[float]],
+        individual_parameters: DictParams,
+        feature: Optional[FeatureType] = None,
+    ) -> torch.Tensor:
+        """
+        For one individual, compute age(s) at which the given features values
+        are reached (given the subject's individual parameters).
+
+        Consistency checks are done in the main :term:`API` layer.
+
+        Parameters
+        ----------
+        value : scalar or array_like[scalar] (:obj:`list`, :obj:`tuple`, :class:`numpy.ndarray`)
+            Contains the :term:`biomarker` value(s) of the subject.
+
+        individual_parameters : :obj:`dict`
+            Contains the individual parameters.
+            Each individual parameter should be a scalar or array_like.
+
+        feature : :obj:`str` (or None)
+            Name of the considered :term:`biomarker`.
+
+            .. note::
+                Optional for :class:`.UnivariateModel`, compulsory
+                for :class:`.MultivariateModel`.
+
+        Returns
+        -------
+        :class:`torch.Tensor`
+            Contains the subject's ages computed at the given values(s).
+            Shape of tensor is ``(1, n_values)``.
+
+        Raises
+        ------
+        :exc:`.LeaspyModelInputError`
+            If computation is tried on more than 1 individual.
+        """
+        # value, individual_parameters = self._get_tensorized_inputs(
+        #     value, individual_parameters, skip_ips_checks=False
+        # )
+        # return self.compute_individual_ages_from_biomarker_values_tensorized(
+        #     value, individual_parameters, feature
+        # )
+        raise NotImplementedError("This method is currently not implemented.")
+
+    def compute_individual_ages_from_biomarker_values_tensorized(
+        self,
+        value: torch.Tensor,
+        individual_parameters: DictParamsTorch,
+        feature: Optional[FeatureType],
+    ) -> torch.Tensor:
+        """
+        For one individual, compute age(s) at which the given features values are
+        reached (given the subject's individual parameters), with tensorized inputs.
+
+        Parameters
+        ----------
+        value : :class:`torch.Tensor` of shape ``(1, n_values)``
+            Contains the :term:`biomarker` value(s) of the subject.
+
+        individual_parameters : DictParamsTorch
+            Contains the individual parameters.
+            Each individual parameter should be a :class:`torch.Tensor`.
+
+        feature : :obj:`str` (or None)
+            Name of the considered :term:`biomarker`.
+
+            .. note::
+                Optional for :class:`.UnivariateModel`, compulsory
+                for :class:`.MultivariateModel`.
+
+        Returns
+        -------
+        :class:`torch.Tensor`
+            Contains the subject's ages computed at the given values(s).
+            Shape of tensor is ``(n_values, 1)``.
+        """
+        raise NotImplementedError("This method is currently not implemented.")
